Revise `ConfigFile` example to be more clear
This pull request addresses the example use-case of the `ConfigFile` shown in the documentation being too unclear/confusing for general use. It has several `if` statements that seem unnecessary in a simple example, and it's more likely to confuse a beginner than to help them understand what a config file is used for.